+++ apps/web/components/surface/modals/index.ts
@@ -1,4 +1,5 @@
-import { createPushModal } from '@zxkit/surface'
+import type React from 'react'
+import { createPushModal, modal } from '@zxkit/surface'
 import { DynamicWrapper } from './dynamic'
 
 import { DefaultModalExample } from './example/default-example'
@@ -6,9 +7,13 @@ import { DynamicModalExample } from './example/dynamic-example'
 import { StateModalExample } from './example/state-example'
 import { FormModalExample } from './example/form-example'
 import { DefaultSheetExample } from './example/sheet-example'
+import { AsyncModalExample } from './example/async-example'
+import { ReplaceStartExample } from './example/replace-start-example'
+import { ReplaceSuccessExample } from './example/replace-success-example'
 
 export const {
   pushModal,
+  pushModalAsync,
   popModal,
   popAllModals,
   replaceWithModal,
@@ -18,23 +23,35 @@ export const {
 } = createPushModal({
   modals: {
     // just dialog
-    DefaultExample: DefaultModalExample,
+    DefaultExample: modal<Record<never, never>>(DefaultModalExample),
 
     // sheet (this doesn't work with dynamic modals)
-    DefaultSheetExample: DefaultSheetExample,
+    DefaultSheetExample: modal<Record<never, never>>(DefaultSheetExample),
 
     // dynamic dialog/drawer based on breakpoint
-    DynamicExample: {
+    DynamicExample: modal<Record<never, never>>({
       Wrapper: DynamicWrapper,
       Component: DynamicModalExample,
-    },
-    StateExample: {
+    }),
+    StateExample: modal<Record<never, never>>({
       Wrapper: DynamicWrapper,
       Component: StateModalExample,
-    },
-    FormExample: {
+    }),
+    FormExample: modal<Record<never, never>>({
       Wrapper: DynamicWrapper,
       Component: FormModalExample,
-    },
+    }),
+    AsyncExample: modal<React.ComponentProps<typeof AsyncModalExample>, boolean>({
+      Wrapper: DynamicWrapper,
+      Component: AsyncModalExample,
+    }),
+    ReplaceStartExample: modal<Record<never, never>>({
+      Wrapper: DynamicWrapper,
+      Component: ReplaceStartExample,
+    }),
+    ReplaceSuccessExample: modal<React.ComponentProps<typeof ReplaceSuccessExample>>({
+      Wrapper: DynamicWrapper,
+      Component: ReplaceSuccessExample,
+    }),
   },
 })